WIP of moving calling to a new window
