+++ ts/components/session/ActionsPanel.tsx
@@ -46,7 +46,7 @@ import { loadDefaultRooms } from '../../opengroup/opengroupV2/ApiUtil';
 
 import { ActionPanelOnionStatusLight } from '../dialog/OnionStatusPathDialog';
 import { switchHtmlToDarkTheme, switchHtmlToLightTheme } from '../../state/ducks/SessionTheme';
-import { CallContainer } from './calling/CallContainer';
+import { IncomingCallContainer } from './calling/IncomingCallContainer';
 const Section = (props: { type: SectionType; avatarPath?: string | null }) => {
   const ourNumber = useSelector(getOurNumber);
   const unreadMessageCount = useSelector(getUnreadMessageCount);
@@ -288,7 +288,7 @@ export const ActionsPanel = () => {
     <>
       <ModalContainer />
 
-      <CallContainer />
+      <IncomingCallContainer />
 
       <div className="module-left-pane__sections-container">
         <Section type={SectionType.Profile} avatarPath={ourPrimaryConversation.avatarPath} />
